Version 23.2.7

+++ dbo.spREN_DeleteTenant.sql
@@ -3,8 +3,8 @@ GO
 SET ANSI_NULLS, QUOTED_IDENTIFIER ON
 GO
 CREATE PROCEDURE [dbo].[spREN_DeleteTenant]
-	@TenantID [bigint] = 0,
-	@UserID [bigint] = 0,
+	@TenantID [int] = 0,
+	@UserID [int] = 0,
 	@RoleID [int] = 0,
 	@LangID [int] = 1
 WITH ENCRYPTION, EXECUTE AS CALLER
@@ -13,7 +13,7 @@ BEGIN TRANSACTION
 BEGIN TRY  
 SET NOCOUNT ON;  
 		--Declaration Section
-		DECLARE @HasAccess bit,@lft bigint,@rgt bigint,@Width bigint
+		DECLARE @HasAccess bit,@lft INT,@rgt INT,@Width INT,@UserName NVARCHAR(64)
 
 		--SP Required Parameters Check
 		if(@TenantID=0)
@@ -38,13 +38,23 @@ SET NOCOUNT ON;
 		SELECT @lft = lft, @rgt = rgt, @Width = rgt - lft + 1
 		FROM REN_Tenant WITH(NOLOCK) WHERE TenantID=@TenantID
 		
-		declare @temp table(id int identity(1,1), TenantID bigint)
+		--ondelete External function
+		IF (@TenantID>0)
+		BEGIN
+			DECLARE @tablename NVARCHAR(200)
+			set @tablename=''
+			select @tablename=SpName from ADM_DocFunctions a WITH(NOLOCK) where CostCenterID=94 and Mode=8
+			if(@tablename<>'')
+				exec @tablename 94,@TenantID,'',@UserID,@LangID	
+		END			
+		
+		declare @temp table(id int identity(1,1), TenantID INT)
 		
 		insert into @temp
 		select DISTINCT TenantID  from REN_Tenant with(nolock) WHERE lft >= @lft AND rgt <= @rgt
 		
 		declare @i int, @cnt int
-		DECLARE @NodeID bigint, @Dimesion bigint 
+		DECLARE @NodeID INT, @Dimesion INT 
 		set @i=1
 		select @cnt=count(*) from @temp
 		 
@@ -60,7 +70,7 @@ SET NOCOUNT ON;
 			begin
 				Update REN_Tenant set CCID=0, CCNodeID=0 where TenantID in
 				(select TenantID from @temp where id=@i)
-				declare @return_value bigint
+				declare @return_value INT
 			  
 				EXEC @return_value = [dbo].[spCOM_DeleteCostCenter]
 					@CostCenterID = @Dimesion,
@@ -73,40 +83,60 @@ SET NOCOUNT ON;
 				--Deleting from Mapping Table
 				Delete from com_docbridge WHERE CostCenterID = 94 AND RefDimensionNodeID = @NodeID AND RefDimensionID = 	@Dimesion					
 			end
+
+			--Delete Linked Account
+				set @NodeID=0
+				select  @NodeID = AccountID from REN_Tenant with(nolock) 
+				where TenantID  IN (select TenantID from @temp where id=@i)
+
+				if (@NodeID is not null and @NodeID>0)
+				begin
+					Update REN_Tenant set AccountID=0 where TenantID in (select TenantID from @temp where id=@i)
+				
+					EXEC spACC_DeleteAccount @NodeID,@UserID,1,@LangID
+				
+					--Deleting from Mapping Table
+					Delete from com_docbridge WHERE CostCenterID = 94 AND RefDimensionNodeID = @NodeID AND RefDimensionID = 2					
+				end
+			--END :: Delete Linked Account
+
+
 			set @i=@i+1
 		end
-
-		INSERT INTO [dbo].[REN_TenantHistory]
-		([TenantID],[TenantCode],[TypeID],[PositionID],[FirstName],[MiddleName],[LastName],[LeaseSignatory],[ContactPerson],[PostingID]
-		,[Phone1],[Phone2],[Email],[Fax],[IDNumber],[Profession],[Passport],[Nationality],[PassportIssueDate],[PassportExpiryDate]
-		,[SponsorName],[SponsorPassport],[SponsorIssueDate],[SponsorExpiryDate],[License],[LicenseIssuedBy],[LicenseIssueDate]
-		,[LicenseExpiryDate],[Description],[Depth],[ParentID],[lft],[rgt],[IsGroup],[CompanyGUID],[GUID],[CreatedBy],[CreatedDate]
-		,[ModifiedBy],[ModifiedDate],[CCNodeID],[CCID],[UserName],[Password],[StatusID],[HistoryStatus])
-		select 
-		[TenantID],[TenantCode],[TypeID],[PositionID],[FirstName],[MiddleName],[LastName],[LeaseSignatory],[ContactPerson],[PostingID]
-		,[Phone1],[Phone2],[Email],[Fax],[IDNumber],[Profession],[Passport],[Nationality],[PassportIssueDate],[PassportExpiryDate]
-		,[SponsorName],[SponsorPassport],[SponsorIssueDate],[SponsorExpiryDate],[License],[LicenseIssuedBy],[LicenseIssueDate]
-		,[LicenseExpiryDate],[Description],[Depth],[ParentID],[lft],[rgt],[IsGroup],[CompanyGUID],[GUID],[CreatedBy],[CreatedDate]
-		,[ModifiedBy],[ModifiedDate],[CCNodeID],[CCID],[UserName],[Password],[StatusID],'Deleted'
-		from ren_tenant with(nolock) WHERE lft >= @lft AND rgt <= @rgt
+		
+		SELECT @UserName=USERNAME FROM ADM_USERS WITH(NOLOCK) WHERE UserID=@UserID
 	
-		insert into Ren_TenantExtendedHistory
-		select *,'Deleted' from REN_TenantExtended WHERE Tenantid IN
-		(select Tenantid from REN_Tenant with(nolock) WHERE lft >= @lft AND rgt <= @rgt)
+		select @i=1,@cnt=count(*) from @temp
+		while @i<=@cnt
+		begin
+			select @NodeID=TenantID from @temp where id=@i
+			
+			--INSERT INTO HISTROY   
+			EXEC [spCOM_SaveHistory]  
+				@CostCenterID =94,    
+				@NodeID =@NodeID,
+				@HistoryStatus ='Deleted',
+				@UserName=@UserName
+			
+			set @i=@i+1
+		end
+			
+		DELETE FROM  COM_Notes  
+		WHERE FEATUREID=94 and  FeaturePK in (select TenantID from @temp)
 		
 		DELETE FROM  COM_Files  
-		WHERE FEATUREID=94 and  FeaturePK in
-		(select TenantID from REN_Tenant with(nolock) WHERE lft >= @lft AND rgt <= @rgt)
+		WHERE FEATUREID=94 and  FeaturePK in (select TenantID from @temp)
 		
-		DELETE FROM CRM_ACTIVITIES WHERE CostCenterID=94 AND NodeID in
-		(select TenantID from REN_Tenant with(nolock) WHERE lft >= @lft AND rgt <= @rgt)
+		DELETE FROM CRM_ACTIVITIES WHERE CostCenterID=94 AND NodeID in (select TenantID from @temp)
+		
+		DELETE FROM  COM_Contacts 
+		WHERE FEATUREID=94 and  FeaturePK in (select TenantID from @temp)
 		
 		---Delete from Extended Table
-     	DELETE FROM REN_TenantExtended WHERE TenantID in
-		(select TenantID from REN_Tenant with(nolock) WHERE lft >= @lft AND rgt <= @rgt)
+     	DELETE FROM REN_TenantExtended WHERE TenantID in (select TenantID from @temp)
 		
 		--Delete from main table
-		DELETE FROM REN_Tenant WHERE lft >= @lft AND rgt <= @rgt
+		DELETE FROM REN_Tenant WHERE TenantID in (select TenantID from @temp)
 
 		--Update left and right extent to set the tree
 		UPDATE REN_Tenant SET rgt = rgt - @Width WHERE rgt > @rgt;
@@ -143,5 +173,4 @@ ROLLBACK TRANSACTION
 SET NOCOUNT OFF  
 RETURN -999   
 END CATCH
-
 GO
